Create a basic game class with a code generator
This is the initial commit. It includes a simple game class which, upon
initialization, generates a 4 digit code to be used in the game. As of
now the code is generated without repeats so that the feedback part of
the game can be simplified. Repeats will be implemented later on.