Make message and writer
In this commit I added a long message to be printed at the beginning of
the game. I also added a method which will slowly print out a message
one letter at a time as if it is being typed out. This method will be
used to print out everything in the game. This commit also layed out
the outline of how the rest of the code will work - now just a few more
methods need to be made.