Fix codes not having any repeats
This commit got rid of the add_num method in the Game class and
put the necessary part of it into the generate_code method because there
was very little code and it fit well in the other method. Before this
commit, the add_num method had still not been allowing duplicates in the
@code variable. That part of the method was deleted when it was moved to
the generate_code.