Implement first simple computer guessing version
This is the first version of the game where you can create a code and
the computer will guess it. Right now the program just generates random
codes and sees if they match perfectly but later the computer's guessing
will be refined. The ComputerSolve module was changed to a class so that
it could use instance variables.