Implement computer algorithm & add module
This version has the computer algorithm fully implemented such that it
will guess the code very consistently. It was implemented inside of the
computer_solve class.

A new module was added to the code which takes care of feedback in the
game, meaning it tells the user or computer how much it guessed
correctly. The module was implemented into both of the classes and
helped DRY up the code a lot.